ZA: Move classes related with organisations in views.py to their own file
