+++ pombola/south_africa/views/__init__.py
@@ -1,4 +1,6 @@
 from .constants import *
 from .geolocalization import *
 from .home import *
+from .organisations import *
 from .views import *
+
